+++ config/config.exs
@@ -19,4 +19,18 @@ if Mix.env() == :dev do
       "format",
       "credo"
     ]
+
+  config :k8s,
+    clusters: %{
+      dev: %{
+        conn: "~/.kube/config",
+        conn_opts: [context: "docker-for-desktop"]
+      }
+    }
+
+  config :bonny,
+    cluster_name: :dev,
+    controllers: [
+      DeploymentEventLogController
+    ]
 end