+++ lib/bonny/application.ex
@@ -2,17 +2,9 @@ defmodule Bonny.Application do
   @moduledoc false
 
   use Application
-
+  @impl true
   def start(_type, _args) do
-    reconciler = Supervisor.child_spec(Bonny.Reconciler, [])
-
-    watchers =
-      Bonny.Config.controllers()
-      |> Enum.map(fn controller ->
-        Supervisor.child_spec({Bonny.Watcher, controller}, id: controller)
-      end)
-
-    children = [reconciler | watchers]
+    children = Bonny.Config.controllers()
     opts = [strategy: :one_for_one, name: Bonny.Supervisor]
     Supervisor.start_link(children, opts)
   end