+++ lib/bonny/config.ex
@@ -108,6 +108,7 @@ defmodule Bonny.Config do
   @doc """
   `K8s.Cluster` name used for this operator. Defaults to `:default`
   """
+  @spec cluster_name() :: atom
   def cluster_name() do
     Application.get_env(:bonny, :cluster_name, :default)
   end