+++ lib/bonny/server.ex
@@ -1,3 +1,0 @@
-defmodule Bonny.Server do
-  @moduledoc false
-end