+++ lib/mix/bonny.ex
@@ -19,6 +19,7 @@ defmodule Mix.Bonny do
 
   Special handling for the path "-" will render to STDOUT
   """
+  @spec render(binary, binary) :: binary
   def render(source, "-"), do: IO.puts(source)
 
   def render(source, target) do
@@ -26,27 +27,32 @@ defmodule Mix.Bonny do
   end
 
   @doc "Get the OTP app name"
+  @spec app_name() :: binary
   def app_name() do
     otp_app()
     |> Atom.to_string()
     |> Macro.camelize()
   end
 
+  @spec app_dir_name() :: binary
   def app_dir_name() do
     Macro.underscore(app_name())
   end
 
+  @spec template(binary) :: binary
   def template(name) do
     template_dir = Application.app_dir(:bonny, ["priv", "templates", "bonny.gen"])
     Path.join(template_dir, name)
   end
 
+  @spec no_umbrella! :: any
   def no_umbrella! do
     if Mix.Project.umbrella?() do
       Mix.raise("mix bonny.gen.* can only be run inside an application directory")
     end
   end
 
+  @spec otp_app :: atom
   defp otp_app() do
     Mix.Project.config() |> Keyword.fetch!(:app)
   end