+++ test/bonny/server/reconciler_test.exs
@@ -4,7 +4,7 @@ defmodule Bonny.Server.ReconcilerTest do
   alias Bonny.Server.Reconciler
 
   defmodule TestReconciler do
-    use Bonny.Server.Reconciler, frequency: 15
+    use Bonny.Server.Reconciler, frequency: 15, client: Bonny.K8sMockClient
 
     @impl true
     def reconcile(pod) do
@@ -13,21 +13,17 @@ defmodule Bonny.Server.ReconcilerTest do
     end
 
     @impl true
-    def resources() do
-      {:ok,
-       [
-         %{"name" => "foo"},
-         %{"name" => "bar"}
-       ]}
+    def reconcile_operation() do
+      K8s.Client.list("reconciler.test/v1", :foos)
     end
   end
 
-  test "schedule/2 sends `:run` after N seconds" do
+  test "schedule/2 sends `:run` after delay" do
     Reconciler.schedule(self(), 1)
     assert_receive :run, 2_000
   end
 
-  describe "run/1" do
+  test "run/1" do
     Agent.start_link(fn -> [] end, name: TestReconcilerCache)
     Reconciler.run(TestReconciler)
     Process.sleep(500)