+++ test/support/whizbang.ex
@@ -22,20 +22,29 @@ defmodule Whizbang do
     Agent.update(__MODULE__, fn events -> [event | events] end)
   end
 
+  @impl true
   def add(evt), do: put({:added, evt})
+  @impl true
   def modify(evt), do: put({:modified, evt})
+  @impl true
   def delete(evt), do: put({:deleted, evt})
+  @impl true
   def reconcile(evt), do: put({:reconciled, evt})
 end
 
 defmodule V1.Whizbang do
   @moduledoc false
   use Bonny.Controller
+  @kind "Whizzo"
   @names %{kind: "Whizzo"}
 
+  @impl true
   def add(_), do: :ok
+  @impl true
   def modify(_), do: :ok
+  @impl true
   def delete(_), do: :ok
+  @impl true
   def reconcile(_), do: :ok
 end
 
@@ -46,7 +55,6 @@ defmodule V2.Whizbang do
   @rule {"apiextensions.k8s.io", ["bar"], ["*"]}
   @rule {"apiextensions.k8s.io", ["foo"], ["*"]}
 
-  @version "v2alpha1"
   @group "kewl.example.io"
   @scope :cluster
   @names %{
@@ -56,9 +64,13 @@ defmodule V2.Whizbang do
     shortNames: ["f", "b", "q"]
   }
 
+  @impl true
   def add(_), do: :ok
+  @impl true
   def modify(_), do: :ok
+  @impl true
   def delete(_), do: :ok
+  @impl true
   def reconcile(_), do: :ok
 end
 
@@ -66,7 +78,7 @@ defmodule V3.Whizbang do
   @moduledoc false
   use Bonny.Controller
 
-  @version "v2alpha1"
+  @version "v3alpha1"
   @group "kewl.example.io"
   @scope :cluster
   @names %{
@@ -83,8 +95,12 @@ defmodule V3.Whizbang do
     }
   ]
 
+  @impl true
   def add(_), do: :ok
+  @impl true
   def modify(_), do: :ok
+  @impl true
   def delete(_), do: :ok
+  @impl true
   def reconcile(_), do: :ok
 end