+++ test/test_helper.exs
@@ -1,3 +1,5 @@
-Bonny.Telemetry.DebugLogger.attach()
+K8s.Sys.Logger.attach()
+Bonny.Sys.Logger.attach()
+
 {:ok, _} = Whizbang.start_link()
 ExUnit.start()